+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * CosmicSynth has been folded into the Macabre2D.Framework project.
 * TileableComponents now use a GridConfiguration that allows them to link up to a GridModule instead of a local grid.
 * Modules can be given names.
+* Components can have selectable modules in the editor.
 
 ### Refactors